+++ CHANGELOG.rst
@@ -35,6 +35,8 @@ Added
   - ``/`` and ``<PREFIX>/`` redirect to ``/<PREFIX>/admin`` (if enabled).
 
 - New validate_config method in GraphQL API.
+- Add ``zone`` and ``zone_distances`` parameters to server and cluster helpers
+  respectively.
 
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 Changed
@@ -44,7 +46,7 @@ Changed
   according to the leader order from topology.
 - Update ``errors`` dependency to 2.2.0.
 - Add default ``pool.map_call`` timeout 10 seconds.
-- Update ``frontend-core`` dependency to 7.8.0.
+- Update ``frontend-core`` dependency to 7.9.1.
 - Argparse throws an error when it encouters ``instance_name`` missing in
   instances.yml.